修改了投票详情接口，包含了用户投票选项
