+++ post-service/post-server/src/main/java/com/xueyu/post/pojo/vo/VoteVO.java
@@ -43,6 +43,11 @@ public class VoteVO {
     Integer amount;
 
     /**
+     * 是否投票 若集合为空则为未投票或未登录，投票则返回用户投票选项id的集合
+     */
+    List<Integer> isVote;
+
+    /**
      * 选项列表
      */
     List<VoteOption> optionList;
