+++ post-service/post-server/src/main/java/com/xueyu/post/service/VoteRecordService.java
@@ -1,6 +1,7 @@
 package com.xueyu.post.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
+import com.xueyu.post.pojo.domain.Vote;
 import com.xueyu.post.pojo.domain.VoteRecord;
 
 import java.util.List;
@@ -25,4 +26,12 @@ public interface VoteRecordService extends IService<VoteRecord> {
      * @return {@link List}<{@link Integer}>
      */
     List<Integer> isVote(Integer userId, Integer voteId);
+
+    /**
+     * 投票是否过期
+     *
+     * @param vote 投票
+     * @return {@link Boolean}
+     */
+    Boolean isVoteExpired(Vote vote);
 }
